サイドバーのコンポーネント化
