+++ front/src/components/molecules/NavItem.tsx
@@ -1,0 +1,53 @@
+import React from "react";
+
+import {
+  Box,
+  Flex,
+  Icon,
+  FlexProps,
+} from '@chakra-ui/react'
+import { IconType } from 'react-icons';
+
+interface NavItemProps extends FlexProps {
+  icon: IconType
+  children: string
+}
+
+const NavItem = ({ icon, children, ...rest }: NavItemProps) => {
+  return (
+    <>
+      <Box
+        as="a"
+        href="#"
+        style={{ textDecoration: 'none' }}
+        _focus={{ boxShadow: 'none' }}>
+        <Flex
+          align="center"
+          p="4"
+          mx="4"
+          borderRadius="lg"
+          role="group"
+          cursor="pointer"
+          _hover={{
+            bg: 'cyan.400',
+            color: 'white',
+          }}
+          {...rest}>
+          {icon && (
+            <Icon
+              mr="4"
+              fontSize="16"
+              _groupHover={{
+                color: 'white',
+              }}
+              as={icon}
+            />
+          )}
+          {children}
+        </Flex>
+      </Box>
+    </>
+  )
+};
+
+export default NavItem;
