+++ front/src/components/organisms/SideBarContent.tsx
@@ -1,0 +1,64 @@
+import React from "react";
+
+import {
+  Box,
+  CloseButton,
+  Flex,
+  useColorModeValue,
+  Text,
+  BoxProps,
+} from '@chakra-ui/react'
+import {
+  FiHome,
+  FiTrendingUp,
+  FiCompass,
+  FiStar,
+  FiSettings,
+} from 'react-icons/fi'
+import { IconType } from 'react-icons';
+import NavItem from "../molecules/NavItem";
+
+interface LinkItemProps {
+  name: string
+  icon: IconType
+}
+
+interface SidebarProps extends BoxProps {
+  onClose: () => void
+}
+
+const LinkItems: Array<LinkItemProps> = [
+  { name: 'Home', icon: FiHome },
+  { name: 'Trending', icon: FiTrendingUp },
+  { name: 'Explore', icon: FiCompass },
+  { name: 'Favourites', icon: FiStar },
+  { name: 'Settings', icon: FiSettings },
+]
+
+const SidebarContent = ({ onClose, ...rest }: SidebarProps) => {
+  return (
+    <Box
+      bg={useColorModeValue('white', 'gray.900')}
+      borderRight="1px"
+      borderRightColor={useColorModeValue('gray.200', 'gray.700')}
+      w={{ base: 'full', md: 'full' }}
+      pos="fixed"
+      h="full"
+      zIndex={10}
+      {...rest}>
+      <Flex h="20" alignItems="center" mx="8" justifyContent="space-between">
+        <Text fontSize="2xl" fontFamily="monospace" fontWeight="bold">
+          画HACK
+        </Text>
+        <CloseButton display={{ base: 'flex', md: 'none' }} onClick={onClose} />
+      </Flex>
+      {LinkItems.map((link) => (
+        <NavItem key={link.name} icon={link.icon}>
+          {link.name}
+        </NavItem>
+      ))}
+    </Box>
+  )
+};
+
+export default SidebarContent;
