+++ front/src/components/pages/TopPage.tsx
@@ -1,18 +1,18 @@
 import React from 'react';
-import { Center } from '@chakra-ui/react';
-import Header from '../templates/Header';
+import { Center, Box, useColorModeValue} from '@chakra-ui/react';
 import Footer from '../templates/Footer';
 import Sidebar from '../templates/SideBar';
 
 function TopPage() {
   return (
     <>
-      <Sidebar />
-      <Header />
-      <Center h={'500px'}>
-        <h1>Hello! Gahack</h1>
-      </Center>
-      <Footer />
+      <Box bg={useColorModeValue('white', 'gray.900')}>
+        <Sidebar />
+        <Center h={'500px'}>
+          <h1>Hello! Gahack</h1>
+        </Center>
+        <Footer />
+      </Box>
     </>
   );
 }
